requirements.txt
requests
flask
pytest
pytest-pycodestyle
datetime
pymodm
dnspython
Pillow
scipy
matplotlib